Support 'manifest' as a gerrit project
Allows to handle changes that include changes in the manifest.

This change adds more explicit control for the manifest
repository. We now support specifying the gerrit server, the
manifest project, as well as the branch in the manifest
repository for 'init'.

NOTE: This works not always as expected because we would need
an additional 'sync' cycle after downloading changes in the
manifest (and these changes could pop up recursively).

Change-Id: Iecfe5d1a9f42a8ebb13c70433ba492ae1ece7614