Add support for --cleanup during sync
The --cleanup option for the sync command enables cleanup the working
tree, this means projects that are no longer in the manifest are moved
into an attic directory. However, only projects with a clean working
tree are moved away, the sync fails otherwise.
Additionally we now support moving projects to different locations with
an updated manifest.

Change-Id: I2fb90a0e5c277117612049bd6f389e6d8bb5bdd1