Add ssh connection checking
This adds checking of the ssh connection because ssh might
request interactive response from the user, such as when
asking for a passphrase (for ssh-agent) or when asking to valide
a new host key. The check is a pragmatic variant of avoiding
the case of ham just hanging and doing nothing when
ssh is requesting something from the user but which is not shown.

Change-Id: I6f93379da6b91fce8159bcf48c5c4302f7ee0e0a