Add additional 'project_name' attribute to a project
Allow additional attribute 'project_name' for project nodes in the
manifest. 'project_name' is equal to 'name' if omitted, and is appended
to the remote URL to locate the project on the remote. This allows to
specify a different project name on the remote server and keeping the
local name unchanged.  This provides consistent local naming even when
migrating moving repositories on their servers.

Change-Id: I142b6b4fca44f277d14305c9454a1610c159cf1e