download-id considers manifest changes
When ham download-id changes projects it now also checks
if those manifest projects registered in the manifest index
file were changed.

With the new option

  ham download-id --sync
  ham download-id --sync=arg1,arg2

triggers a

  ham sync

with provided arguments turned into ham sync options, e.g.,

  ham sync --arg1 --arg2

after the patches are applied, so that changes to manifests
are reflected in the projects list, i.e., new projects are
checked out or deleted projects are moved into the attic. Usual
options could be 'cleanup' or 'force-local-update' or both.

Change-Id: I775fd446856c261a24f1e6a59b440a8cdd9607bb