ham - Exit with code 1 in ham download-id if no change was found
ham download-id does not return with a non-zero exit code if it does
not find any open Gerrit changes for its projects to download with the
specified id. This can lead to false positives during automated
testing as the resulting sources do not contain the change to be
tested and the test result does not reflect if the change works or
not.

Change-Id: If52d6ffee23594d77d9743ba85935ed806376858