forall: add {PRJPATH} replacement
I noticed that using git-worktree with a ham-managed repository
structure is quite cumbersome.  This patch tries to address one of
several issues, viz. that there's no pragmatic way to determine the
current repository besides analysing the current working directory.

This change adds a further replacement {PRJPATH} for -i/--vars option
that will be replaced by the respective project's local path (as in the
manifest):

  $ ham forall l4re-core -i -- echo 'pn: {PRJNAME}; pp: {PRJPATH}'
  pn: l4re-core; pp: pkg/l4re-core

In theory, a worktree can then be created with e.g.

  l4root $ git worktree add --detach /path/to/wt # [1]
  l4root $ ham forall -i -- git worktree add --detach \
                  '/path/to/wt/{PRJPATH}'
  l4root $ ln -s ${PWD}/.ham /path/to/wt/
  l4root $ cd /path/to/wt/
  l4root $ ham checkout some-branch

(checkout will fail due to other problems not addressed here)

[1] unnecessary if repo belonging to . was first in manifest

Change-Id: I9bed8daba534c9cfe4e560780d2385f5be0c1036