Fix download-id --print-footers issue
With this change the download-id --print-footers code does now
explicitly and exclusively use the remote branch of the current
repository to generate the list of downloaded commits.

This fixes an issue where a commit is part of any unrelated remote
branch already and therefore excluded from the list of downloaded
commits even if not merged officially to its branch.

Change-Id: I191751c2d6285aaf8e029551091fabc7825e0d63