sync: Add --exclude-manifest
To be able to use earlier revisions of the manifest repositories for
the purpose of reproducing previous source setups, we want the sync
command to not update the manifest repositories during the first sync
after init. To achieve this we add the --exclude-manifest flag to the
sync command which skips projects marked with is_manifest during sync.

Change-Id: I428252a37da071568ae738479fd23fed4543f4c8