Add detailed log of operations to download-id
Print a log of operations done during download-id.
A log line has the form:
log: <project-name>: <project-path>: <fetch-url>: <operation> <revision>

where operation is checkout, cherry-pick, or merge.
The operations are in the same order as applied by ham, and the
revisions are the revisions found on the given remote system.

Change-Id: I1f638a5c6aa2a574e3975f5df2e8e3fd7f8f08ae