Set is_manifest when loading local manifest projects
When saving local manifests into the local manifest index the
is_manifest flag is not saved and it is not added again either when
loading the projects from the local manifest index at a later point in
time. This leads to the local manifests repositories not being treated
like manifest repos after ham init. To fix this we add the is_manifest
flag again when loading projects from the local manifest index.

Change-Id: I7b11179e6504e179bae05af056b966ead606a6b2