prevent checkout into detached HEAD state being treated as error
When git checkouts a tag it does so creating a detached HEAD. git also
prints a note to stderr informing the user about the potential to shoot
himself into the foot. Unfortunately ham considers this to be an error
and prints the hint as an error message and also returns with an error
code.

The checkout command already filters a couple of messages. This change
now does the same for this situation. ham still prints the git message
but as info now. ham also returns with the actual result from git
checkout.

Change-Id: Ic3fcf24c417987b2041f2949e5abfd3d80f7589b