ham - Replace XML::Mini::Document with XML::Parser
The XML::Mini::Document module does not parse xml comments correctly. It
just treats <!-- --> as a differently named tag which only works as long
as the comment does not contain a ">", which however is the case when
commenting out other tags such as a project tag. By shear coincidence it
works out perfectly when commenting out only one project tag, because it
treats everything from the comment's "<" until the project's ">" as one
tag with the name "!--", which is ignored by ham. The remaining --> from
the comment is then treated as text data. When commenting out two
projects in one comment the same happens with the first project tag, but
the second commented out project tag is treated as if it is not
commented out at all.

Since this is a bug within the XML::Mini::Document module that cannot be
fixed in ham, we decide to replace the module with the XML::Parser
module, which handles comments correctly and only needs a bit of code to
generate the same data structure we got from XML::Mini::Document.

Change-Id: Ibb34513a525ecad3204292721f6aa339cf48762a