ham - Present correct revision of downloaded change
After downloading a change ham listed the changes and revisions
downloaded in the log. For this ham looked at the HEAD commit after
downloading. This works when using checkout and maybe merge for applying
the downloaded change, but not if the change is added to the repository
using --rebase or when it's cherry picked. To be sure we list the
original revision, we have to look at FETCH_HEAD instead of HEAD.

Change-Id: I75ec747ec4694dd23e0a3a8f94267c9ffba39cf8