init: Create manifest for default revision
This change adds code that creates a separate manifest containing a
default revision whenever the --branch parameter is provided. This
default revision will then apply to all subsequent ham commands run in
the same ham directory.

Furthermore we add the possibility for the user to specify the
default-revision themself. This is especially useful for cases where we
don't have a branch parameter.

Change-Id: I64f84262b1ef89a81bf5d82735572ebffe7a8af3