ham - Give helpful warning if branch doesn't exist in local manifest
When using a local manifest repository which does not have the expected
branch, ham shows a warning that the local manifest file is not present
this repository. This is misleading because the underlying issue is the
branch being missing and thus the worktree stays empty. This change adds
a separate warning in case the branch is missing.

Change-Id: Id02a390814c5de449229f9a4687a4b31884dcb64