ham - Add -C option similar to git
git has a -C option which causes the program to change to a specified
directory before continuing execution. This is especially helpful if you
want to do git operations in a different directory than the current work
directory. This change implements this option for ham. In this
implementation "-C" has to be the first argument to ham, followed by the
target path.

Change-Id: I7caa4cf44853eca2730aec0ea68ab44d62268967