ham - Handle transition for manifest url format change
Previously the manifests_json only contained the base path of the remote
without the projects name. We decided to change that in
I141cece4514394dea463e2b798532665c19d6774. The transitional code however
turned out to be erroneous and was removed in
Ic1017622643d127cdfd034937614d8ad32997928. However we still need some
sort of transition mechanism for ham trees that have not been updated in
a while, so we add a file version to the manifests_repo so we know this
tree has already run through that transition.

The same is true for local manifests. Newly created local-manifest
projects have a dedicated remote entry, which contains the complete url.
Older local-manifest project do not have that. In that case, their fetch
url needs to be handled like in non-manifest repos.

Change-Id: Ib1d29986666ffb922335f4e626c13f6c6a27429d