ham - Use Hammer::Manifest to generate LMI during ham init
With the Manifest loading outsourced into the Hammer::Manifest perl
module, we can also migrate the Manifest saving to this perl module.

For this we add code to Hammer::Manifest which generates XML. Ham
generates XML by itself in an effort to prevent any new dependencies.

Furthermore we start using this code by generating the Local Manifest
Index with it during ham init.

Change-Id: I1e70e78ac10638c2a309af538d4714a7b893ce47