ham/init: Update default revision on reinit
A while ago we introduced a default revision manifest which
automatically sets the initially specified branch of the manifest
project as default revision for all other projects. Up until now this
default revision was not updated when running ham init again with a
different branch. This change fixes this by recreating that default
revision manifest.

Change-Id: I6f5e3a3f1299b5405e858aacb52882a29d3400cb