ham/init: Do not unconditionally pin local-manifest revision in LMI
During the work to deduplicate local manifests we unconditionally wrote
the revision into the local manifest index, no matter if it was
selected explicitly for that local-manifest or if that revision came
from the --branch parameter as a fallback. For the most part this makes
no difference until you run ham init again to change the manifest branch
and the default revision, in which case the fixed revision in the local
manifest index still applies.

To fix this we have to differentiate between a local-manifest repo that
is supposed to have revision X explicitly and a local-manifest repo that
is supposed to have revision X because revision X was selected for the
main manifest repository.

Change-Id: Ied121528e89ee85c1d3c0a81d452fccb62b0dce5