ham: Do not implicitly use manifest branch as default for --branch
Up until now an absent --branch and --revision lead to ham determining
the --branch value automatically from the cloned manifest repository, so
that it doesn't matter to other initialization steps whether the branch
was specified or not. In case of --default-revision which defaults to
the --branch value this is an undesired outcome, because the default
revision manifest shall not be created if --branch was not specified.

Change-Id: I60d7d422739981cfe0c62f088e44df1f9cb871e5