Hammer::Project: Explicitly fetch configured revision
Up until a while ago ham relied on a simple `git fetch <remote>`
fetching everything necessary to check out a particular revision. Since
then we added commit hashes as an exception which have to be explicitly
retrieved. However this approach doesn't work for tags that are not
present on any branches. To fix this we explicitly fetch the selected
revision (if applicable) and also request all remote branch refs to be
updated, which mimics the former default behavior.

Solves issue kernkonzept/ham#7

Signed-off-by: Valentin Gehrke <valentin.gehrke@kernkonzept.com>
Change-Id: Ia049e93c5430dcb7a30daf9a37c00d3216b4556b