ham: Sanity check for multiple projects using same path
Up until now ham allowed several projects to use the same path. This
leads to ham managing these projects in the same .git dir and only
checking out one of them as a worktree. It's unlikely there is a
use-case for this and actually can cause confusion. To prevent this we
add a sanity check that makes sure that cannot happen.

Change-Id: I22e269972df5102f735ab9fd49c367a0cbba9984
Signed-off-by: Valentin Gehrke <valentin.gehrke@kernkonzept.com>